MAINT: Begin modularizing tests
